+++ lib/src/layer/tile_layer.dart
@@ -960,7 +960,9 @@ class _TileLayerState extends State<TileLayer> with TickerProviderStateMixin {
       );
     }
 
-    setState(() {});
+    if (mounted) {
+      setState(() {});
+    }
 
     if (_noTilesToLoad()) {
       // Wait a bit more than tileFadeInDuration (the duration of the tile fade-in)
@@ -970,7 +972,9 @@ class _TileLayerState extends State<TileLayer> with TickerProviderStateMixin {
             ? options.tileFadeInDuration + const Duration(milliseconds: 50)
             : const Duration(milliseconds: 50),
         () {
-          setState(_pruneTiles);
+          if (mounted) {
+            setState(_pruneTiles);
+          }
         },
       );
     }
@@ -1082,6 +1086,7 @@ class Tile implements Comparable<Tile> {
     }
 
     animationController?.removeStatusListener(_onAnimateEnd);
+    animationController?.dispose();
     _imageStream?.removeListener(_listener);
   }
 
@@ -1197,7 +1202,9 @@ class _AnimatedTileState extends State<AnimatedTile> {
   }
 
   void _handleChange() {
-    setState(() {});
+    if (mounted) {
+      setState(() {});
+    }
   }
 }